pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "propicks"
version = "0.1.0"
description = "Trading engine AI-driven: scoring tecnico, sizing, journal, report"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Valerio Uberti" }]
license = { text = "Proprietary" }
dependencies = [
    "yfinance>=0.2.40",
    "pandas>=2.0",
    "numpy>=1.24",
    "tabulate>=0.9",
    "anthropic>=0.40",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.8",
]

[project.scripts]
propicks-scan      = "propicks.cli.scanner:main"
propicks-portfolio = "propicks.cli.portfolio:main"
propicks-journal   = "propicks.cli.journal:main"
propicks-report    = "propicks.cli.report:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
